chore: add review suggestions
