+++ src/app/components/confirmBtcTransaction/itemRow/amount.tsx
@@ -1,17 +1,16 @@
 import TokenImage from '@components/tokenImage';
 import useWalletSelector from '@hooks/useWalletSelector';
 import { currencySymbolMap, getBtcFiatEquivalent, satsToBtc } from '@secretkeylabs/xverse-core';
+import Avatar from '@ui-library/avatar';
 import { StyledP } from '@ui-library/common.styled';
 import BigNumber from 'bignumber.js';
-import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { NumericFormat } from 'react-number-format';
 import styled from 'styled-components';
-import Avatar from './avatar';
 
-export interface TokenImageProps {
+type Props = {
   amount: number;
-}
+};
 
 const RowCenter = styled.div<{ spaceBetween?: boolean }>((props) => ({
   display: 'flex',
@@ -29,11 +28,15 @@ const AvatarContainer = styled.div`
   margin-right: ${(props) => props.theme.space.xs};
 `;
 
-export default function Amount({ amount }: TokenImageProps) {
+export default function Amount({ amount }: Props) {
   const { btcFiatRate, fiatCurrency } = useWalletSelector();
   const { t } = useTranslation('translation');
 
-  const getFiatAmountString = (fiatAmount: BigNumber) => {
+  const getFiatAmountString = (amountParam: number, btcFiatRateParam: string) => {
+    const fiatAmount = getBtcFiatEquivalent(
+      new BigNumber(amountParam),
+      BigNumber(btcFiatRateParam),
+    );
     if (!fiatAmount) {
       return '';
     }
@@ -74,9 +77,7 @@ export default function Amount({ amount }: TokenImageProps) {
             renderText={(value: string) => <StyledP typography="body_medium_m">{value}</StyledP>}
           />
           <StyledP typography="body_medium_s" color="white_400">
-            {getFiatAmountString(
-              getBtcFiatEquivalent(new BigNumber(amount), BigNumber(btcFiatRate)),
-            )}
+            {getFiatAmountString(amount, btcFiatRate)}
           </StyledP>
         </NumberTypeContainer>
       </RowCenter>
